pyproject.toml
[project]
name = "zoro"
version = "0.1.0"
description = ""
authors = [
    {name = "Yoru-cyber",email = "carlosmendez170210@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "customtkinter (>=5.2.2,<6.0.0)",
    "pillow (>=11.1.0,<12.0.0)"
]

[tool.poetry]
packages = [{include = "zoro", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
